Fix bug in code for varargs functions with multiple fixed parameters.
This was broken by the varargs changes in commit 4e15176bec5d5. The code was not accounting for the internal representation of the parameters being in reverse order, so it was basing address calculations on the first fixed parameter rather than the last one, resulting in the wrong number of bytes being removed from the stack (generally causing a crash).

This affected the c99stdio.c test case, and is now also covered in c99stdarg.c.