Fix varargs stack repair.
Varargs-only stack repair (i.e. using #pragma optimize bit 3 but not bit 6) was broken by commit aa12e155a332b3a. It removed some code that was needed to allocate the direct page location used to hold the stack pointer value in that case. This would lead to invalid code being produced, which could cause a crash when run. The fix is to revert the erroneous parts of commit aa12e155a332b3a (which do not affect its core purpose of enabling intermediate code peephole optimization to be used when stack repair code is active).