Update strtod to follow C99+ requirements.
Specifically, this adds support for:
*Numbers in hexadecimal floating-point format
*"INFINITY" ("INF" was already recognized)
*"NAN(n-char-sequence)" (this was already recognized if it used digits only)

Also, line feed, form feed, carriage return, and vertical tab characters can now be skipped as part of the leading whitespace. (This was required even in C90.)

Additionally, it now sets errno to EINVAL if no conversion can be performed because the string did not have a suitable format. This is not required by Standard C, but is required by POSIX.