printf: ensure %g prints 0 in a non-exponential format.
It could sometimes print 0 as 0e+00, depending on previous operations. This could occur because the exp value is undefined after an FX2DEC operation on 0 (see Apple Numerics Manual 2nd Ed. p. 28). In practice, the old value seems to be left unmodified, and it can vary based on the previous conversion operation.

Here is a program that exhibited this issue in my testing:

#include <stdio.h>
int main(void) {
        printf("%g\n", 1.0);
        printf("%g\n", 0.0);
}