Fix bug in display of local variables when DP is $8000 or higher.
Local variable values would not be displayed correctly in this case, because the 16-bit DP value would be treated as signed and sign-extended to 32 bits. If the DP value was $8000 or greater, the bank byte would be set to $FF rather than $00, so Prizm would read the values for the variables from the wrong location. The fix is to ensure that the DP value is always zero-extended.