Parse type names involving function types more correctly.
This affects type names such as "int (int)", which should indicate a function type. Previously, it would not be parsed as such.

Function types are not generally legal where type names are used (other than in typeof, but that's not yet implemented), so use of such type names will still result in an error, but a better error message may be reported.

This also removes the previous special-case code to handle type names like "int ()". They can now be handled by the general code for function abstract declarators.

This still does not actually process prototypes in type names to generate the appropriate prototyped function type. That will be handled later.