Add support for variable length array types.
This follows the C23 requirement to support VLA types, but not necessarily to support local variables of VLA type (which we still do not). This should be essentially a full implementation of VLA types as specified by C23, except that errors about branching into the scope of a variably modified variable are not detected.