Declare unsigned comparison functions as returning a 16-bit integer.
This is consistent with what the assembly code actually does. Previously, the result was treated as a 32-bit integer, but since the assembly code only returns 16 bits, the upper 16 bits would be garbage. This wouldn't end up mattering, because the result was later truncated to 16 bits, but it's cleaner to just declare the functions correctly.