Implement _BitInt types (C23).
This is a near-complete implementation of _BitInt types as specified in the C23 standard. The only missing element is support for _BitInt bitfields wider than 32 bits.

This implementation supports _BitInt widths up to 64 and stores _BitInt types using either 16, 32, or 64 bits, whichever is required to fit them. They are sign-extended or zero-extended out to the width used to store them, and for calling conventions and most other purposes they are treated like integers of that width. Unsigned _BitInt types are truncated and zero-extended as necessary to implement the standard wrap-around semantics for unsigned types. When a value is converted to a signed _BitInt type, it is truncated and sign-extended. However, signed _BitInt types are generally not truncated and sign-extended in cases of arithmetic overflow, since that would be fairly expensive and overflow is UB according to the standard; thus, overflow involving signed _BitInt types can effectively produce values outside the range of the type.

This also incidentally (due to changes in GetLongLongVal) fixes a bug where constant division operations such as -1 / 12345u could be mis-evaluated because the usual arithmetic conversions were not applied to them correctly. (Both operands would be promoted to wider types, rather than just converting -1 to 0xFFFF.)